refactor(root): change app structure to follow auth workflow
wrap base app under a provider store to allow global auth based routing
